workflows/gen_graph/agents/user_intent/pyproject.toml
[project]
name = "user-intent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "dagger-io",
    "ais-dagger-agents-config>=0.1.0,<0.2.0",
    "pyyaml>=6.0.2",
    "pydantic[email]>=2.11.7",
    "pydantic-ai>=1.0.1",
    "simple-chalk>=0.1.0",
    "uvicorn>=0.28.0",
    "fastapi>=0.110.0",
    "typing-extensions>=4.12.0",
    "starlette>=0.36.3",
]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }
